requirements/common.txt
requests==2.23.0
python-bitcoinlib==0.11.0
uncertainties==3.1.2
flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Script==2.0.6
flask-restplus==0.13.0
Flask-Testing==0.8.0
Werkzeug==0.16.1
